HSEARCH-3324 Split BridgeIT into four separate tests
It was becoming too large, and it addressed multiple different features.
